# server/api/gen/go/config/config.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.11
// 	protoc        (unknown)
// source: config/config.proto

package config

import (
	_ "github.com/google/gnostic/openapiv3"
	common "github.com/liujitcn/shop-base/server/api/gen/go/common"
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type ConfigRequest struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Site          common.BaseConfigSite  `protobuf:"varint,2,opt,name=site,proto3,enum=common.BaseConfigSite" json:"site,omitempty"` // 位置：枚举【BaseConfigSite】
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ConfigRequest) Reset() {
	*x = ConfigRequest{}
	mi := &file_config_config_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ConfigRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ConfigRequest) ProtoMessage() {}

func (x *ConfigRequest) ProtoReflect() protoreflect.Message {
	mi := &file_config_config_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ConfigRequest.ProtoReflect.Descriptor instead.
func (*ConfigRequest) Descriptor() ([]byte, []int) {
	return file_config_config_proto_rawDescGZIP(), []int{0}
}

func (x *ConfigRequest) GetSite() common.BaseConfigSite {
	if x != nil {
		return x.Site
	}
	return common.BaseConfigSite(0)
}

type ConfigResponse struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Data          []*ConfigResponse_Data `protobuf:"bytes,1,rep,name=data,proto3" json:"data,omitempty"` // 系统配置
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ConfigResponse) Reset() {
	*x = ConfigResponse{}
	mi := &file_config_config_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ConfigResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ConfigResponse) ProtoMessage() {}

func (x *ConfigResponse) ProtoReflect() protoreflect.Message {
	mi := &file_config_config_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ConfigResponse.ProtoReflect.Descriptor instead.
func (*ConfigResponse) Descriptor() ([]byte, []int) {
	return file_config_config_proto_rawDescGZIP(), []int{1}
}

func (x *ConfigResponse) GetData() []*ConfigResponse_Data {
	if x != nil {
		return x.Data
	}
	return nil
}

type ConfigResponse_Data struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Key           string                 `protobuf:"bytes,1,opt,name=key,proto3" json:"key,omitempty"`     // 配置key
	Value         string                 `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"` // 配置value
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ConfigResponse_Data) Reset() {
	*x = ConfigResponse_Data{}
	mi := &file_config_config_proto_msgTypes[2]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ConfigResponse_Data) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ConfigResponse_Data) ProtoMessage() {}

func (x *ConfigResponse_Data) ProtoReflect() protoreflect.Message {
	mi := &file_config_config_proto_msgTypes[2]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ConfigResponse_Data.ProtoReflect.Descriptor instead.
func (*ConfigResponse_Data) Descriptor() ([]byte, []int) {
	return file_config_config_proto_rawDescGZIP(), []int{1, 0}
}

func (x *ConfigResponse_Data) GetKey() string {
	if x != nil {
		return x.Key
	}
	return ""
}

func (x *ConfigResponse_Data) GetValue() string {
	if x != nil {
		return x.Value
	}
	return ""
}

var File_config_config_proto protoreflect.FileDescriptor

const file_config_config_proto_rawDesc = "" +
	"\n" +
	"\x13config/config.proto\x12\x06config\x1a$gnostic/openapi/v3/annotations.proto\x1a\x1cgoogle/api/annotations.proto\x1a\x11common/enum.proto\"f\n" +
	"\rConfigRequest\x12U\n" +
	"\x04site\x18\x02 \x01(\x0e2\x16.common.BaseConfigSiteB)\xbaG&\x92\x02#位置：枚举【BaseConfigSite】R\x04site\"\xa9\x01\n" +
	"\x0eConfigResponse\x12C\n" +
	"\x04data\x18\x01 \x03(\v2\x1b.config.ConfigResponse.DataB\x12\xbaG\x0f\x92\x02\f系统配置R\x04data\x1aR\n" +
	"\x04Data\x12!\n" +
	"\x03key\x18\x01 \x01(\tB\x0f\xbaG\f\x92\x02\t配置keyR\x03key\x12'\n" +
	"\x05value\x18\x02 \x01(\tB\x11\xbaG\x0e\x92\x02\v配置valueR\x05value2`\n" +
	"\rConfigService\x12O\n" +
	"\tGetConfig\x12\x15.config.ConfigRequest\x1a\x16.config.ConfigResponse\"\x13\x82\xd3\xe4\x93\x02\r\x12\v/api/configB\x89\x01\n" +
	"\n" +
	"com.configB\vConfigProtoP\x01Z6github.com/liujitcn/shop-base/server/api/gen/go/config\xa2\x02\x03CXX\xaa\x02\x06Config\xca\x02\x06Config\xe2\x02\x12Config\\GPBMetadata\xea\x02\x06Configb\x06proto3"

var (
	file_config_config_proto_rawDescOnce sync.Once
	file_config_config_proto_rawDescData []byte
)

func file_config_config_proto_rawDescGZIP() []byte {
	file_config_config_proto_rawDescOnce.Do(func() {
		file_config_config_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_config_config_proto_rawDesc), len(file_config_config_proto_rawDesc)))
	})
	return file_config_config_proto_rawDescData
}

var file_config_config_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
var file_config_config_proto_goTypes = []any{
	(*ConfigRequest)(nil),       // 0: config.ConfigRequest
	(*ConfigResponse)(nil),      // 1: config.ConfigResponse
	(*ConfigResponse_Data)(nil), // 2: config.ConfigResponse.Data
	(common.BaseConfigSite)(0),  // 3: common.BaseConfigSite
}
var file_config_config_proto_depIdxs = []int32{
	3, // 0: config.ConfigRequest.site:type_name -> common.BaseConfigSite
	2, // 1: config.ConfigResponse.data:type_name -> config.ConfigResponse.Data
	0, // 2: config.ConfigService.GetConfig:input_type -> config.ConfigRequest
	1, // 3: config.ConfigService.GetConfig:output_type -> config.ConfigResponse
	3, // [3:4] is the sub-list for method output_type
	2, // [2:3] is the sub-list for method input_type
	2, // [2:2] is the sub-list for extension type_name
	2, // [2:2] is the sub-list for extension extendee
	0, // [0:2] is the sub-list for field type_name
}

func init() { file_config_config_proto_init() }
func file_config_config_proto_init() {
	if File_config_config_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_config_config_proto_rawDesc), len(file_config_config_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   3,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_config_config_proto_goTypes,
		DependencyIndexes: file_config_config_proto_depIdxs,
		MessageInfos:      file_config_config_proto_msgTypes,
	}.Build()
	File_config_config_proto = out.File
	file_config_config_proto_goTypes = nil
	file_config_config_proto_depIdxs = nil
}